Use bridge as single source of truth for events
